Change cluster_limit in config to an Option<usize>
This was originally a u16 (defaulting to 1) which is more similar to what the old backend config parsing used, but now if it's None tokio can automatically detect the number of worker threads to use based on the number of CPUs.

+++ packages/backend/crates/server/src/lib.rs
@@ -19,7 +19,7 @@ pub fn init() -> Result<(), Box<dyn error::Error>> {
     let rt = rt.enable_all();
 
     if let Some(n) = get_config()?.cluster_limit {
-        rt.worker_threads(n as usize);
+        rt.worker_threads(n);
     }
 
     let rt = rt.build()?;
